# client/api/openapi.yaml
openapi: 3.0.2
info:
  contact:
    url: https://github.com/moov-io/watchman
  description: Moov Watchman is an HTTP API and Go library to download, parse and
    offer search functions over numerous trade sanction lists from the United States,
    European Union governments, agencies, and non profits for complying with regional
    laws. Also included is a web UI and async webhook notification service to initiate
    processes on remote systems.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Watchman API
  version: v1
servers:
- description: Local development
  url: http://localhost:8084
...
paths:
...
  /ofac/companies/{companyID}:
    get:
      description: Get information about a company, trust or organization such as
        addresses, alternate names, and remarks.
      operationId: getOfacCompany
      parameters:
      - description: Optional Request ID allows application developer to trace requests
          through the systems logs
        explode: false
        in: header
        name: X-Request-ID
        required: false
        schema:
          example: 94c825ee
          type: string
        style: simple
      - description: Company ID
        explode: false
        in: path
        name: companyID
        required: true
        schema:
          example: 1d1c824a
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OfacCompany'
          description: Company and associated metadata
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Error occured, see response body.
      summary: Get company
      tags:
      - Watchman
    put:
      description: Update a Companies sanction status to always block or always allow
        transactions.
      operationId: updateOfacCompanyStatus
      parameters:
      - description: Optional Request ID allows application developer to trace requests
          through the systems logs
        explode: false
        in: header
        name: X-Request-ID
        required: false
        schema:
          example: 94c825ee
          type: string
        style: simple
      - description: User ID used to perform this search
        explode: false
        in: header
        name: X-User-ID
        required: false
        schema:
          type: string
        style: simple
      - description: Company ID
        explode: false
        in: path
        name: companyID
        required: true
        schema:
          example: 8d49fd61
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateOfacCompanyStatus'
        required: true
      responses:
        "200":
          description: Company status updated
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Error occured, see response body.
      summary: Update company
      tags:
      - Watchman
  /ofac/companies/{companyID}/watch:
    post:
      description: Add name watch on a OFAC Company
      operationId: addOfacCompanyWatch
      parameters:
      - description: Optional Request ID allows application developer to trace requests
          through the systems logs
        explode: false
        in: header
        name: X-Request-ID
        required: false
        schema:
          example: 94c825ee
          type: string
        style: simple
      - description: Company ID
        explode: false
        in: path
        name: companyID
        required: true
        schema:
          example: c3cf0f66
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OfacWatchRequest'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OfacWatch'
          description: Company watch added
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Error occured, see response body.
        "404":
          description: Company not found
      summary: Watch OFAC company
      tags:
      - Watchman
  /ofac/companies/{companyID}/watch/{watchID}:
    delete:
      description: Delete a company name watch
      operationId: removeOfacCompanyWatch
      parameters:
      - description: Optional Request ID allows application developer to trace requests
          through the systems logs
        explode: false
        in: header
        name: X-Request-ID
        required: false
        schema:
          example: 94c825ee
          type: string
        style: simple
      - description: Company ID
        explode: false
        in: path
        name: companyID
        required: true
        schema:
          example: c3cf0f66
          type: string
        style: simple
      - description: Watch ID, used to identify a specific watch
        explode: false
        in: path
        name: watchID
        required: true
        schema:
          example: 0c5e215c
          type: string
        style: simple
      responses:
        "200":
          description: Company watch removed
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Error occured, see response body.
      summary: Remove company watch
      tags:
      - Watchman
  /ofac/companies/watch:
    post:
      description: Watch a company by its name. The match percentage will be included
        in the webhook's JSON payload.
      operationId: addOfacCompanyNameWatch
      parameters:
      - description: Optional Request ID allows application developer to trace requests
          through the systems logs
        explode: false
        in: header
        name: X-Request-ID
        required: false
        schema:
          example: 94c825ee
          type: string
        style: simple
      - description: Company name used to match and send watch notifications
        explode: true
        in: query
        name: name
        required: true
        schema:
          example: Jane Smith
          type: string
        style: form
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OfacWatchRequest'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OfacWatch'
          description: Company watch applied
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Error occured, see response body.
      summary: Watch company
      tags:
      - Watchman
  /ofac/companies/watch/{watchID}:
    delete:
      operationId: removeOfacCompanyNameWatch
      parameters:
      - description: Optional Request ID allows application developer to trace requests
          through the systems logs
        explode: false
        in: header
        name: X-Request-ID
        required: false
        schema:
          example: 94c825ee
          type: string
        style: simple
      - description: Watch ID, used to identify a specific watch
        explode: false
        in: path
        name: watchID
        required: true
        schema:
          example: 0c5e215c
          type: string
        style: simple
      - description: Company name watch
        explode: true
        in: query
        name: name
        required: true
        schema:
          example: Jane Smith
          type: string
        style: form
      responses:
        "200":
          description: Company watch removed
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Error occured, see response body.
      summary: Remove company watch
      tags:
      - Watchman
...
components:
  schemas:
    OfacCompany:
      description: OFAC Company and metadata
      example:
        addresses:
        - country: Japan
          address: 123 73th St
          cityStateProvincePostalCode: Tokyo 103
          match: 0.91
          entityID: "2112"
          addressID: "201"
        - country: Japan
          address: 123 73th St
          cityStateProvincePostalCode: Tokyo 103
          match: 0.91
          entityID: "2112"
          addressID: "201"
        alts:
        - alternateID: "220"
          alternateRemarks: Extra information
          alternateType: aka
          match: 0.91
          entityID: "306"
          alternateName: NATIONAL BANK OF CUBA
        - alternateID: "220"
          alternateRemarks: Extra information
          alternateType: aka
          match: 0.91
          entityID: "306"
          alternateName: NATIONAL BANK OF CUBA
        ID: 9574756b
        sdn:
          match: 0.91
          entityID: "1231"
          programs:
          - CUBA
          sdnName: BANCO NACIONAL DE CUBA
          title: Title of an individual
          remarks: remarks
        status:
          note: Incorrect match
          createdAt: 2000-01-23T04:56:07.000+00:00
          userID: 349661f9
          status: unsafe
      properties:
        ID:
          description: OFAC Company ID
          example: 9574756b
          type: string
        sdn:
          $ref: '#/components/schemas/OfacSDN'
        addresses:
          items:
            $ref: '#/components/schemas/OfacEntityAddress'
          type: array
        alts:
          items:
            $ref: '#/components/schemas/OfacAlt'
          type: array
        status:
          $ref: '#/components/schemas/OfacCompanyStatus'
    OfacCompanyStatus:
      description: Status properties of an OFAC Company
      example:
        note: Incorrect match
        createdAt: 2000-01-23T04:56:07.000+00:00
        userID: 349661f9
        status: unsafe
      properties:
        userID:
          description: User ID provided when updating status
          example: 349661f9
          type: string
        note:
          description: Optional note from updating status
          example: Incorrect match
          type: string
        status:
          description: Manually applied status for OFAC Company
          enum:
          - unsafe
          - exception
          type: string
        createdAt:
          format: date-time
          type: string
...
    OfacSDN:
      description: Specially designated national from OFAC list
      example:
        match: 0.91
        entityID: "1231"
        programs:
        - CUBA
        sdnName: BANCO NACIONAL DE CUBA
        title: Title of an individual
        remarks: remarks
      properties:
        entityID:
          example: "1231"
          type: string
        sdnName:
          example: BANCO NACIONAL DE CUBA
          type: string
        sdnType:
          $ref: '#/components/schemas/SdnType'
        programs:
          description: Programs is the sanction programs this SDN was added from
          example:
          - CUBA
          items:
            type: string
          type: array
        title:
          example: Title of an individual
          type: string
        remarks:
          type: string
        match:
          description: Remarks on SDN and often additional information about the SDN
          example: 0.91
          type: number
...
    OfacEntityAddress:
      description: Physical address from OFAC list
      example:
        country: Japan
        address: 123 73th St
        cityStateProvincePostalCode: Tokyo 103
        match: 0.91
        entityID: "2112"
        addressID: "201"
      properties:
        entityID:
          example: "2112"
          type: string
        addressID:
          example: "201"
          type: string
        address:
          example: 123 73th St
          type: string
        cityStateProvincePostalCode:
          example: Tokyo 103
          type: string
        country:
          example: Japan
          type: string
        match:
          example: 0.91
          type: number
...
    OfacAlt:
      description: Alternate name from OFAC list
      example:
        alternateID: "220"
        alternateRemarks: Extra information
        alternateType: aka
        match: 0.91
        entityID: "306"
        alternateName: NATIONAL BANK OF CUBA
      properties:
        entityID:
          example: "306"
          type: string
        alternateID:
          example: "220"
          type: string
        alternateType:
          example: aka
          type: string
        alternateName:
          example: NATIONAL BANK OF CUBA
          type: string
        alternateRemarks:
          example: Extra information
          type: string
        match:
          example: 0.91
          type: number
    SdnType:
      description: Used for classifying SDNs — typically represents an individual
        or company
      enum:
      - individual
      - entity
      - vessel
      - aircraft
      type: string
...
    UpdateOfacCompanyStatus:
      description: Request body to update a company status.
      example:
        notes: False positive
        status: unsafe
      properties:
        status:
          description: manual override of company/SDN sanction status
          enum:
          - unsafe
          - exception
          type: string
        notes:
          description: Free form notes about manually changing the Company status
          example: False positive
          type: string
      required:
      - status
...
    OfacWatch:
      description: Customer or Company watch
      example:
        watchID: 08ddba92
      properties:
        watchID:
          description: Object representing a customer or company watch
          example: 08ddba92
          type: string
    OfacWatchRequest:
      description: Webhook or other means of notification on search criteria. OFAC
        will make a POST request with a body of the customer or company (SDN, AltNames,
        and Address).
      example:
        webhook: https://api.example.com/ofac/webhook
        authToken: 75d0384b-a105-4048-9fce-91a280ce7337
      properties:
        authToken:
          description: Private token supplied by clients to be used for authenticating
            webhooks.
          example: 75d0384b-a105-4048-9fce-91a280ce7337
          type: string
        webhook:
          description: HTTPS url for webhook on search match
          example: https://api.example.com/ofac/webhook
          type: string
      required:
      - authToken
      - webhook
...
    Error:
      properties:
        error:
          description: An error message describing the problem intended for humans.
          example: Example error, see description
          type: string
      required:
      - error